sort by estimated hours, sort function now dry, removed main and hooked up layout instead, markers cleared from map on submit of new partload and global map instance created...work out how to export mapObject!!

+++ react/client/components/account_management/UserList.jsx
@@ -36,6 +36,11 @@ class UserList extends React.Component {
   handleDeleteClick(e){
     e.preventDefault()
     var id = e.target.id
+    console.log('id' , id, this.props.users.currentUser.id)
+    if(id==this.props.users.currentUser.id){
+      alert('Current User cannot be deleted!')
+      return
+    }
     this.props.deleteUser(id)
   }
